pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "SurfaceAlpha"
version = "0.1.0"
description = "Multimodal volatility surface + returns modeling for regime-conditional forecasting and tradable portfolio overlay"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Ethan Bobrik"},
]

dependencies = [
    # ── Core scientific stack ──
    "numpy>=1.24,<2.0",
    "pandas>=2.0",
    "scipy>=1.10",
    "scikit-learn>=1.3",

    # ── Deep learning ──
    "torch>=2.1",
    "timm>=0.9",                  # Vision Transformer building blocks

    # ── Econometric baselines ──
    "arch>=6.0",                  # GARCH family
    "statsmodels>=0.14",          # HAR-RV, OLS

    # ── Shallow ML baselines ──
    "lightgbm>=4.0",
    "xgboost>=2.0",

    # ── Explainability ──
    "shap>=0.42",

    # ── Data sources ──
    "yfinance>=0.2.30",           # Underlying OHLCV + VIX
    "doltcli>=0.1.18",            # DoltHub option_chain + volatility_history

    # ── Configuration ──
    "hydra-core>=1.3",
    "omegaconf>=2.3",

    # ── Experiment tracking ──
    "wandb>=0.16",

    # ── Visualization ──
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "plotly>=5.15",

    # ── Utilities ──
    "pyarrow>=14.0",              # Parquet I/O
    "tqdm>=4.65",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
    "ipykernel>=6.25",
    "jupyter>=1.0",
    "nbformat>=5.9",
]

[project.scripts]
sa-fetch-options = "scripts.fetch_options:main"
sa-fetch-underlying = "scripts.fetch_underlying:main"
sa-fetch-market = "scripts.fetch_market_state:main"
sa-build-canonical = "scripts.build_canonical:main"
sa-build-surfaces = "scripts.build_surfaces:main"
sa-train = "scripts.train:main"
sa-evaluate = "scripts.evaluate:main"
sa-backtest = "scripts.backtest:main"
sa-explain = "scripts.explain:main"

[tool.setuptools.packages.find]
where = ["src"]

# ── Pytest ──
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests that require external data sources",
]

# ── Ruff (linter + formatter) ──
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
]
ignore = [
    "E501",  # line too long — handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["volregime"]

# ── Mypy ──
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false          # relaxed for research code
ignore_missing_imports = true          # many scientific libs lack stubs
